requirements_pi.txt
#tflite_runtime-1.14.0-cp37-cp37m-linux_aarch64.whl on PC
tflite_runtime-1.14.0-cp37-cp37m-linux_armv7l.whl
Adafruit_blinka==3.7.1
Adafruit_circuitpython_BME280==2.3.3
pyaudio==0.2.11
requests==2.21.0
numpy==1.16.4
soundfile==0.10.2
watchdog==0.9.0
datetime==4.3
librosa==0.7.2
coverage==5.0.3

